+++ docs/src/shared/ComponentPreviewTabs.tsx
@@ -4,11 +4,10 @@ import Tab from '@mui/material/Tab';
 import Box from '@mui/material/Box';
 import { Link, Outlet, useLocation, useNavigate } from 'react-router-dom';
 import { SxProps, Theme, useTheme } from '@mui/material/styles';
-import { DRAWER_WIDTH, TabPanel } from '../shared';
+import { TabPanel } from '../shared';
 import { PLAYGROUND_DRAWER_WIDTH } from './constants';
 
-const hidePlaygroudTabs = ['drawer-layout', 'spacer', 'drawer-body'];
-const docsTabs = ['examples', 'api-docs', 'playground'];
+const hidePlaygroundTabs = ['drawer-layout', 'spacer', 'drawer-body'];
 
 function a11yProps(index: number): any {
     return {
@@ -17,11 +16,11 @@ function a11yProps(index: number): any {
     };
 }
 
-function getNumber(location: string, tabs: string[]): number {
-    const pathname = tabs.includes(location.split('/')[4]) ? location.split('/')[4] : location.split('/')[3];
-    if (!pathname) return 0;
+function getTabNumber(location: string): number {
+    const pathName = location.substring(location.lastIndexOf('/') + 1);
+    if (!pathName) return 0;
 
-    switch (pathname) {
+    switch (pathName) {
         case 'api-docs':
             return 1;
         case 'playground':
@@ -32,8 +31,8 @@ function getNumber(location: string, tabs: string[]): number {
 }
 
 function togglePlaygroundTab(location: string): boolean {
-    const tabName = location.split('/').filter((e) => hidePlaygroudTabs.includes(e))[0];
-    return hidePlaygroudTabs.includes(tabName);
+    const tabName = location.split('/').filter((e) => hidePlaygroundTabs.includes(e))[0];
+    return hidePlaygroundTabs.includes(tabName);
 }
 
 const tabStyles = {
@@ -54,7 +53,7 @@ const tabPanelContentStyles: SxProps<Theme> = {
 };
 
 const outletContainerStyles = {
-    p: '48px 0',
+    pb: 6,
 };
 
 const playgroundContentStyles = {
@@ -70,47 +69,68 @@ export const ComponentPreviewTabs = (): JSX.Element => {
     const location = useLocation();
     const [value, setValue] = React.useState(0);
     const [hidePlaygroundTab, setHidePlaygroundTab] = React.useState(false);
+
     const theme = useTheme();
     const handleChange = (event: React.SyntheticEvent, newValue: number): void => {
         navigate(`/${newValue === 1 ? 'api-docs' : newValue === 2 ? 'playground' : 'examples'}`);
     };
 
     React.useEffect(() => {
-        setValue(getNumber(location?.pathname, docsTabs));
+        setValue(getTabNumber(location?.pathname));
         setHidePlaygroundTab(togglePlaygroundTab(location.pathname));
     }, [location]);
 
     return (
-        <Box sx={{ width: '100%' }}>
-            <Box
+        <>
+            <Tabs
+                value={value}
+                onChange={handleChange}
+                aria-label="component docs tabs"
+                centered
                 sx={{
+                    width: '100%',
+                    display: 'flex',
+                    justifyContent: 'space-evenly',
                     bgcolor: theme.palette.background.paper,
                     borderBottom: 1,
                     borderColor: 'divider',
-                    width: `calc(100% - ${DRAWER_WIDTH}px)`,
-                    position: 'fixed',
+                    position: 'sticky',
+                    top: { xs: 56, sm: 64 },
                     zIndex: theme.zIndex.appBar,
-                    [theme.breakpoints.down('lg')]: {
-                        width: '100%',
+                    '& .MuiTabs-indicator': {
+                        backgroundColor: 'primary.main',
                     },
                 }}
             >
-                <Tabs
-                    value={value}
-                    onChange={handleChange}
-                    aria-label="component docs tabs"
-                    centered
+                <Tab
+                    to="examples"
+                    component={Link}
                     sx={{
-                        width: '100%',
-                        display: 'flex',
-                        justifyContent: 'space-evenly',
-                        '& .MuiTabs-indicator': {
-                            backgroundColor: 'primary.main',
+                        ...tabStyles,
+                        [theme.breakpoints.down(1040)]: {
+                            width: '33%',
                         },
                     }}
-                >
+                    label="Examples"
+                    replace={true}
+                    {...a11yProps(0)}
+                />
+                <Tab
+                    to="api-docs"
+                    component={Link}
+                    sx={{
+                        ...tabStyles,
+                        [theme.breakpoints.down(1040)]: {
+                            width: '33%',
+                        },
+                    }}
+                    label="API Docs"
+                    replace={true}
+                    {...a11yProps(1)}
+                />
+                {!hidePlaygroundTab && (
                     <Tab
-                        to="examples"
+                        to="playground"
                         component={Link}
                         sx={{
                             ...tabStyles,
@@ -118,40 +138,12 @@ export const ComponentPreviewTabs = (): JSX.Element => {
                                 width: '33%',
                             },
                         }}
-                        label="Examples"
+                        label="Playground"
                         replace={true}
-                        {...a11yProps(0)}
+                        {...a11yProps(2)}
                     />
-                    <Tab
-                        to="api-docs"
-                        component={Link}
-                        sx={{
-                            ...tabStyles,
-                            [theme.breakpoints.down(1040)]: {
-                                width: '33%',
-                            },
-                        }}
-                        label="API Docs"
-                        replace={true}
-                        {...a11yProps(1)}
-                    />
-                    {!hidePlaygroundTab && (
-                        <Tab
-                            to="playground"
-                            component={Link}
-                            sx={{
-                                ...tabStyles,
-                                [theme.breakpoints.down(1040)]: {
-                                    width: '33%',
-                                },
-                            }}
-                            label="Playground"
-                            replace={true}
-                            {...a11yProps(2)}
-                        />
-                    )}
-                </Tabs>
-            </Box>
+                )}
+            </Tabs>
             <TabPanel value={value} index={0}>
                 <Box
                     sx={{
@@ -187,6 +179,6 @@ export const ComponentPreviewTabs = (): JSX.Element => {
                     <Outlet />
                 </Box>
             </TabPanel>
-        </Box>
+        </>
     );
 };